moving primary gui script output of main window dir
